Refactor To-MimeKit methods to extension methods
