+++ tests/NHSD.BuyingCatalogue.Organisations.Api.UnitTests/Services/EmailAddressTests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using FluentAssertions;
-using MimeKit;
 using NHSD.BuyingCatalogue.Organisations.Api.Services;
 using NUnit.Framework;
 
@@ -37,20 +36,6 @@ namespace NHSD.BuyingCatalogue.Organisations.Api.UnitTests.Services
         public void Address_Set_NullAddress_ThrowsException()
         {
             Assert.Throws<ArgumentNullException>(() => new EmailAddress { Address = null });
-        }
-
-        [Test]
-        public void AsMailboxAddress_ReturnsExpectedValue()
-        {
-            const string name = "Some Body";
-            const string address = "somebody@notarealaddress.co.uk";
-
-            var emailAddress = new EmailAddress(name, address);
-            var mailboxAddress = emailAddress.AsMailboxAddress();
-
-            mailboxAddress.Should().BeOfType<MailboxAddress>();
-            mailboxAddress.Name.Should().Be(name);
-            mailboxAddress.Address.Should().Be(address);
         }
     }
 }
